+++ packages/sources/cryptocompare-test/src/crypto-utils.ts
@@ -1,3 +1,8 @@
+import { SingleNumberResultResponse } from '@chainlink/external-adapter-framework/util'
+import { SettingsMap } from '@chainlink/external-adapter-framework/config'
+import { defaultEndpoint } from './config'
+import { PriceEndpointParams } from '@chainlink/external-adapter-framework/adapter'
+
 export const cryptoInputParams = {
   base: {
     aliases: ['from', 'coin', 'fsym'],
@@ -12,3 +17,159 @@ export const cryptoInputParams = {
     type: 'string',
   },
 } as const
+
+export const endpoints = ['crypto-ws', 'price', 'volume', 'marketcap']
+export const endpointOptions = [defaultEndpoint, ...endpoints]
+
+export const cryptoEndpointInputParams = {
+  ...cryptoInputParams,
+  endpoint: {
+    default: defaultEndpoint,
+    options: endpointOptions,
+    type: 'string',
+  },
+} as const
+
+export type CryptoEndpointParams = PriceEndpointParams & {
+  endpoint: string
+}
+
+export interface ProviderCryptoQuoteData {
+  TYPE: string
+  MARKET: string
+  FROMSYMBOL: string
+  TOSYMBOL: string
+  FLAGS: string
+  PRICE: number
+  LASTUPDATE: number
+  MEDIAN: number
+  LASTVOLUME: number
+  LASTVOLUMETO: number
+  LASTTRADEID: string
+  VOLUMEDAY: number
+  VOLUMEDAYTO: number
+  VOLUME24HOUR: number
+  VOLUME24HOURTO: number
+  OPENDAY: number
+  HIGHDAY: number
+  LOWDAY: number
+  OPEN24HOUR: number
+  HIGH24HOUR: number
+  LOW24HOUR: number
+  LASTMARKET: string
+  VOLUMEHOUR: number
+  VOLUMEHOURTO: number
+  OPENHOUR: number
+  HIGHHOUR: number
+  LOWHOUR: number
+  TOPTIERVOLUME24HOUR: number
+  TOPTIERVOLUME24HOURTO: number
+  CHANGE24HOUR: number
+  CHANGEPCT24HOUR: number
+  CHANGEDAY: number
+  CHANGEPCTDAY: number
+  CHANGEHOUR: number
+  CHANGEPCTHOUR: number
+  CONVERSIONTYPE: string
+  CONVERSIONSYMBOL: string
+  SUPPLY: number
+  MKTCAP: number
+  MKTCAPPENALTY: number
+  TOTALVOLUME24H: number
+  TOTALVOLUME24HTO: number
+  TOTALTOPTIERVOLUME24H: number
+  TOTALTOPTIERVOLUME24HTO: number
+  IMAGEURL: string
+}
+
+export interface ProviderCryptoResponseBody {
+  RAW: {
+    [fsym: string]: {
+      [tsym: string]: ProviderCryptoQuoteData
+    }
+  }
+  DISPLAY: {
+    [fsym: string]: {
+      [tsym: string]: {
+        FROMSYMBOL: string
+        TOSYMBOL: string
+        MARKET: string
+        PRICE: string
+        LASTUPDATE: string
+        LASTVOLUME: string
+        LASTVOLUMETO: string
+        LASTTRADEID: string
+        VOLUMEDAY: string
+        VOLUMEDAYTO: string
+        VOLUME24HOUR: string
+        VOLUME24HOURTO: string
+        OPENDAY: string
+        HIGHDAY: string
+        LOWDAY: string
+        OPEN24HOUR: string
+        HIGH24HOUR: string
+        LOW24HOUR: string
+        LASTMARKET: string
+        VOLUMEHOUR: string
+        VOLUMEHOURTO: string
+        OPENHOUR: string
+        HIGHHOUR: string
+        LOWHOUR: string
+        TOPTIERVOLUME24HOUR: string
+        TOPTIERVOLUME24HOURTO: string
+        CHANGE24HOUR: string
+        CHANGEPCT24HOUR: string
+        CHANGEDAY: string
+        CHANGEPCTDAY: string
+        CHANGEHOUR: string
+        CHANGEPCTHOUR: string
+        CONVERSIONTYPE: string
+        CONVERSIONSYMBOL: string
+        SUPPLY: string
+        MKTCAP: string
+        MKTCAPPENALTY: string
+        TOTALVOLUME24H: string
+        TOTALVOLUME24HTO: string
+        TOTALTOPTIERVOLUME24H: string
+        TOTALTOPTIERVOLUME24HTO: string
+        IMAGEURL: string
+      }
+    }
+  }
+}
+
+interface WSSuccessType {
+  PRICE?: number // Cryptocompare does not provide the price in updates from all exchanges
+  TYPE: string
+  MARKET: string
+  FLAGS: number
+  FROMSYMBOL: string
+  TOSYMBOL: string
+  VOLUMEDAY: number
+  VOLUME24HOUR: number
+  VOLUMEDAYTO: number
+  VOLUME24HOURTO: number
+  MESSAGE?: string
+}
+
+interface WSErrorType {
+  TYPE: string
+  MESSAGE: string
+  PARAMETER: string
+  INFO: string
+}
+
+export type WsMessage = WSSuccessType | WSErrorType
+
+export type CryptoEndpointTypes = {
+  Request: {
+    Params: CryptoEndpointParams
+  }
+  Response: SingleNumberResultResponse
+  CustomSettings: SettingsMap
+  Provider: {
+    RequestBody: ProviderCryptoResponseBody
+    ResponseBody: ProviderCryptoResponseBody
+    WsMessage: WsMessage
+  }
+}
