+++ packages/sources/cryptocompare-test/src/endpoint/crypto-ws.ts
@@ -1,55 +1,14 @@
-import { PriceEndpoint } from '@chainlink/external-adapter-framework/adapter'
 import {
   WebSocketTransport,
   WebSocketRawData,
 } from '@chainlink/external-adapter-framework/transports/websocket'
-import {
-  ProviderResult,
-  makeLogger,
-  SingleNumberResultResponse,
-} from '@chainlink/external-adapter-framework/util'
+import { ProviderResult, makeLogger } from '@chainlink/external-adapter-framework/util'
 import { DEFAULT_WS_API_ENDPOINT } from '../config'
-import { cryptoInputParams } from '../crypto-utils'
-import { CryptoEndpointParams } from './crypto'
-import { SettingsMap } from '@chainlink/external-adapter-framework/config'
-
-interface WSSuccessType {
-  PRICE?: number // Cryptocompare does not provide the price in updates from all exchanges
-  TYPE: string
-  MARKET: string
-  FLAGS: number
-  FROMSYMBOL: string
-  TOSYMBOL: string
-  VOLUMEDAY: number
-  VOLUME24HOUR: number
-  VOLUMEDAYTO: number
-  VOLUME24HOURTO: number
-  MESSAGE?: string
-}
-
-interface WSErrorType {
-  TYPE: string
-  MESSAGE: string
-  PARAMETER: string
-  INFO: string
-}
-
-export type WsMessage = WSSuccessType | WSErrorType
+import { CryptoEndpointTypes } from '../crypto-utils'
 
 const logger = makeLogger('CryptoCompareCryptoEndpoint')
 
-type CryptoWsEndpointTypes = {
-  Request: {
-    Params: CryptoEndpointParams
-  }
-  Response: SingleNumberResultResponse
-  CustomSettings: SettingsMap
-  Provider: {
-    WsMessage: WsMessage
-  }
-}
-
-export const transport = new WebSocketTransport<CryptoWsEndpointTypes>({
+export const wsTransport = new WebSocketTransport<CryptoEndpointTypes>({
   url: (context) => `${DEFAULT_WS_API_ENDPOINT}?api_key=${context.adapterConfig.API_KEY}`,
   handlers: {
     open(connection) {
@@ -66,12 +25,12 @@ export const transport = new WebSocketTransport<CryptoWsEndpointTypes>({
         })
       })
     },
-    message(message): ProviderResult<CryptoWsEndpointTypes>[] | undefined {
+    message(message): ProviderResult<CryptoEndpointTypes>[] | undefined {
       logger.trace(message, 'Got response from websocket')
       if (message.TYPE === '5' && 'PRICE' in message) {
         return [
           {
-            params: { base: message.FROMSYMBOL, quote: message.TOSYMBOL },
+            params: { base: message.FROMSYMBOL, quote: message.TOSYMBOL, endpoint: 'crypto-ws' },
             value: message.PRICE as number,
           },
         ]
@@ -92,10 +51,4 @@ export const transport = new WebSocketTransport<CryptoWsEndpointTypes>({
       return { action: 'SubRemove', subs: [`5~CCCAGG~${params.base}~${params.quote}`] }
     },
   },
-})
-
-export const endpoint = new PriceEndpoint({
-  name: 'crypto-ws',
-  transport,
-  inputParameters: cryptoInputParams,
 })
