+++ packages/sources/cryptocompare-test/src/endpoint/index.ts
@@ -1,2 +1,1 @@
-export { endpoint as cryptoWsEndpoint } from './crypto-ws'
 export { endpoint as cryptoEndpoint } from './crypto'
