+++ packages/sources/cryptocompare-test/src/index.ts
@@ -1,6 +1,6 @@
 import { expose } from '@chainlink/external-adapter-framework'
 import { PriceAdapter } from '@chainlink/external-adapter-framework/adapter'
-import { cryptoEndpoint, cryptoWsEndpoint } from './endpoint'
+import { cryptoEndpoint } from './endpoint'
 import { SettingsMap } from '@chainlink/external-adapter-framework/config'
 import { defaultEndpoint } from './config'
 import overrides from './config/overrides.json'
@@ -8,7 +8,7 @@ import overrides from './config/overrides.json'
 export const adapter = new PriceAdapter<SettingsMap>({
   name: 'CRYPTOCOMPARE',
   defaultEndpoint,
-  endpoints: [cryptoEndpoint, cryptoWsEndpoint],
+  endpoints: [cryptoEndpoint],
   rateLimiting: {
     tiers: {
       free: {
